Revert back to using trailing slashes in blog urls
Gatsby worked like a PWA, and so when deploying to run on nginx,
no trailing slash on blog posts worked out of the box.

But Astro is a true SSG, in the sense that it creates index pages
for each blog post folder, so when deploying to run on nginx,
it uses trailing slashes.

It's much easier to just support trailing slashes than not,
so I'll be using trailing slashes on blog urls.

There is one issue with switching to trailing slashes though...
isso comments db will no longer find a blog post's comments,
because it matches exactly on the url (which has changed.)
Unfortunately, it looks like there's no setting in isso to ignore
comments, so I'll have to update the urls in the isso db.
Fortunately, it looks like someone else has done this with a python
script, so I should be able to figure it out.